Vim/Zsh: Update repo / customize theme
This theme is a slight modification of the robbyrussell theme that ships
with zsh. It improves visual parsing when scrolling and shows the full
directory path. This theme script needs to be added via `source` after
oh-my-zsh init, so a documentation update will be provided later when I
have time.